chore: split func
